requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
alembic==1.13.1
openai>=1.30.0
